refactor: dispatches Tranaction validation to separate modules
