+++ apps/omg/lib/omg/state/transaction/validator/fee_claim.ex
@@ -1,0 +1,45 @@
+# Copyright 2019-2020 OmiseGO Pte Ltd
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+defmodule OMG.State.Transaction.Validator.FeeClaim do
+  @moduledoc """
+  Contains generic validation rules for `FeeTokenClaim` transactions. Specific transaction type's validation
+  is passed to `Transaction.Protocol.can_apply?`
+  """
+
+  alias OMG.State.Core
+  alias OMG.State.Transaction
+
+  @type fee_claim_error :: :claiming_unsupported_token | :claiming_more_than_collected
+
+  @spec can_claim_fees(Core.t(), Transaction.Recovered.t()) ::
+          {:ok, :claim_fees, %{}} | {{:error, fee_claim_error()}, Core.t()}
+  def can_claim_fees(
+        %Core{fee_claimer_address: owner, fees_paid: fees_paid} = state,
+        %Transaction.Recovered{signed_tx: %{raw_tx: fee_tx}}
+      ) do
+    with outputs = make_outputs(owner, fees_paid),
+         {:ok, _} <- Transaction.Protocol.can_apply?(fee_tx, outputs) do
+      {:ok, :claim_fees, %{}}
+    else
+      {:error, _reason} = error -> {error, state}
+    end
+  end
+
+  defp make_outputs(owner, fees_paid) do
+    Enum.map(fees_paid, fn {currency, amount} ->
+      Transaction.FeeTokenClaim.new_output(owner, currency, amount)
+    end)
+  end
+end
